@@ -1,13 +1,14 @@
 import 'package:factor_flutter_mobile/core/my_app.dart';
 import 'package:factor_flutter_mobile/firebase_options.dart';
 import 'package:factor_flutter_mobile/models/factor_view_model/hive/factor_view_model_hive.dart';
+import 'package:factor_flutter_mobile/models/store/hive/store_item_view_model_hive.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-import 'package:parse_server_sdk_flutter/parse_server_sdk.dart';
+import 'package:parse_server_sdk_flutter/parse_server_sdk_flutter.dart';
 
 import 'core/fire_base/fcm/firebase_config.dart';
 
@@ -23,6 +24,7 @@ Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Hive.initFlutter();
   Hive.registerAdapter(FactorHomeViewModelHiveAdapter());
+  Hive.registerAdapter(StoreItemViewModelHiveAdapter());
 
   const keyApplicationId = 'nZLXEWExCdEBnZpQ06NzKpF8bUmDpdbqUi39BOxO';
   const keyClientKey = 'SrZA2mfwiAlY7wot7NWw7a2ls31eHTrDiA0NglTw';
