@@ -13,7 +13,6 @@ import 'package:factor_flutter_mobile/views/shared/widgets/factor_app_bar.dart';
 import 'package:factor_flutter_mobile/views/shared/widgets/factor_button.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:iran_appmarket/iran_appmarket.dart';
 
 class FactorBasePage extends GetView<FactorBaseController> {
   FactorBasePage({Key? key}) : super(key: key);
@@ -68,68 +67,13 @@ class FactorBasePage extends GetView<FactorBaseController> {
           ),
           body: SafeArea(
             bottom: false,
-            child: Column(
-              children: [
-                if(controller.currentIndex.value ==0)
-                CarouselSlider.builder(
-                  itemCount: 2,
-                  itemBuilder: (BuildContext context, int itemIndex, int pageViewIndex) =>
-                      Container(
-                        child: Container(
-                            decoration: BoxDecoration(
-                            color: itemIndex ==1? Colors.brown.shade400: Colors.lightBlue.shade400,
-                              borderRadius: BorderRadius.circular(12)
-                            ),
-                            margin: EdgeInsets.all(20),height: 230,
-                        padding: EdgeInsets.all(12),
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.start,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                             Row(
-                                children: [
-                                   Text(
-                                     itemIndex ==1?'کافه پر': 'تسک پر',
-                                    style: TextStyle(color: Colors.white,fontSize: 24),
-                                  ),
-                              // Constants.mediumHorizontalSpacer,
-                                  if(itemIndex ==0)
-                                  Transform.scale(
-                                    scale: 2,
-                                    child: Image.asset(
-                                      taskParLogo,
-                                      width: 50,
-                                      height: 50,
-                                      // fit: BoxFit.cover,
-                                    ),
-                                  )
-                                ],
-                              ),
-                            Constants.mediumVerticalSpacer,
-                            Text(
-                              itemIndex ==1?'انواع قهوه و لوازم قهوه با کافی پر ارسال رایگان ': 'با اپلیکیشن تسک پر به راحتی تسک بزن',
-                                style: TextStyle(color: Colors.white,fontSize: 14),
+            child: SingleChildScrollView(
+              child: Column(
+                children: [
 
-                            ),
-                            Constants.mediumVerticalSpacer,
-
-                            FactorButton(onPressed: (){
-                              // IranAppMarket.showDeveloperApps(AppMarket.cafeBazaar, '568414623994');
-                              IranAppMarket.showDeveloperApps(
-                                  AppMarket.myket, 'com.example.task_par');
-
-                            },
-                              borderColor: Colors.white,
-                              textColor: Colors.white,
-                              titleButton: 'نصب',
-                            ),
-                          ],
-                        ),
-                        ),
-                      ), options: CarouselOptions(),
-                ),
-                _getPageList().elementAt(controller.currentIndex.value),
-              ],
+                  _getPageList().elementAt(controller.currentIndex.value),
+                ],
+              ),
             ),
           ),
           floatingActionButton: Padding(
@@ -144,7 +88,7 @@ class FactorBasePage extends GetView<FactorBaseController> {
                 // FocusManager.instance.primaryFocus?.unfocus();
                 Get.toNamed(FactorRoutes.listTypeFactor,
                     arguments: const ListTypeFactorPage().arguments(
-                        factorHomeList: controller.factorHomeListHive,
+
                         adsViewModel: controller.adsViewModel,
                         isLoadingAd: controller.isLoadingAd));
               },
@@ -163,9 +107,7 @@ class FactorBasePage extends GetView<FactorBaseController> {
 
   List<Widget> _getPageList() {
     return [
-      HomeFactorPage(
-          factorHomeListHiveSearch: controller.factorHomeListHiveSearch,
-          factorHomeListHive: controller.factorHomeListHive),
+      const HomeFactorPage(),
       const MorePage(),
     ];
   }
